backend/requirements.txt
# --- Web Server ---
fastapi==0.110.0
uvicorn==0.27.1
# pydantic==2.6.3
python-dotenv==1.0.1

# --- AI & Agent Logic ---
langchain-groq==0.1.0
langgraph==0.0.30
langchain-community==0.0.28
langchain==0.1.12

# --- Database (MongoDB) ---
# We use 'motor' for asynchronous MongoDB support with FastAPI
motor==3.3.2
pymongo==4.6.2

# --- Tools & Utilities ---
# For your search and wiki tools
duckduckgo-search==5.3.0
wikipedia==1.4.0
requests==2.31.0

